Refactor achievement stuff
All functions for achievements that can be used by *any* page, and do
not necessarily rely on the structure of the main page, are now in their
 own file so it can be imported.
