+++ index.html
@@ -10,6 +10,7 @@
   <script defer type="module" src="script.js"></script>
   <script defer type="module" src="achievements.js"></script>
   <script defer type="module" src="portfolio.js"></script>
+  <script defer type="module" src="achievementsPane.js"></script>
 
   <meta name="author" content="Leandru Martin"/>
   <meta name="color-scheme" content="dark"/>
